Move frequently changed runner params to model file
We often change:
- The learning rate scheluder parameters (this is a hyperparameter)
- The number of epochs (for convenience)
- Checkpointing and logging directories (from experiment to experiment)

Rather than hardcoding these into the runner we've made them parameters
in the model file. This results in:

- Less repetitive code, since we abstract repeated substrings (e.g.
  version) into variables.
- A single file (the model file) containing all frequently updated
  runner-related hyperparameters.

Thus now we're closer to our goal of only tracking changes in a single
file across experiments. Next we need to abstract away modifiable parts
of the generator and augmenter and move them to the model file.

We still need to make the train runner callable so we can change the
model file ran.