Make custom AugmentationList class
The augmentations for GoogLeNet require being shuffled based on
the paper. We implemented a custom class as a wrapper for the list
of custom augmentations. That class provides a uniform interface
for getting the list of augmentations. Furthermore, the getter
method for getting the augmentation list shuffles the augmentations
based on how the AugmentationList class is instantiated. This allows
us to write the generalized framework for running imagenet models
with minimal changes.