Remove augmentations, metrics, and layers specific to imagenet
This file originally specified GoogLeNet. However for our RL
application we will not need augmentations and we will not
likely use accuracy as a performance metric. As such we
eliminated these from this file.

We removed all of the layers so we can specify a model
custom to this application. We also left notes in this file
for other changes needed to make this model fit with what
was done in the DQN paper.