Remove augmentations and generator instance files
In the RL papers we read augmentations were not used.
As such we have removed them altogether from this framework.
Furthermore, the generator for a particular agent will be
specific to the agent so we do not need a shared generator
instance file. Hence we are removing it.

Note we removed albumentations imports across all files
to eliminate any augmentations altogether.