Strip down base sequence class to barebones
Most of our custom sequence class for imagenet consisted of
logic to support augmentations. Since we won't be using
augmentations for RL projects we deleted most of the methods
in the class. That in turn eliminated the need for most of
the parameters in the constructor so we deleted or replaced
most of them.

What remains is the basic interface needed for a custom
sequence class along with some comments on our thoughts
and concerns about how to implement them.