First draft of an environment sequence class
We decided that implementing a concrete sequence
class based on some assumptions about an environment
and model was a good starting point. We can always
abstract further as we develop more.

In this case we assumed a discrete action space and
a DQN type model. We attempted to separate different
functions of the sequence class in such a way that when
we want to implement double DQN, A3C, etc. we can use
the same interface but change the method code.

At a high level this sequence class separates the tasks of...

- Simulating the environment
- Calculating values and TD errors at specific timesteps
- Managing replay memory (or generally memory of recent states)
- Generating minibatches
- Updating the target model

...into different methods. This way if we wanted to change the
sequence to use for Atari games instead of classic control we
really only need to change the value calculation and minibatch
generation code.

As it is this class is specifically designed for classical
control RL. It does not stack states as is done in DQN for
Atari games. The extension however should be straightforward.
This is a proof of concept for our technical approach, which we
will test with classical control environments.