Create custom callback to evaluate RL agent
Conventional validation as done in Keras does not apply to
the RL use case we have implemented. This is because in
RL we are less interested in the loss between the value
function and actual reward, than we are interested in
the performance of the agent itself. We can measure
the performance of the agent in two ways:

- Selected random states, calculate the value function on
  them, take their average. We expect that as the agent
  learns a better policy the expected value of states
  should increase overall.
- Run the agent for N episodes and average the total
  collected rewards across episodes.

Both of these were used in the original Atari RL papers.

In order to implement this we created a callback that
runs simulations using the agent after an epoch ends.
The callback logs the results of the simulations and
of sampling initial states to the Tensorboard. The callback
uses the shared agent decision methods we put into the
shared methods file.