Fix action selection from Q model
The model predicting Q values expects an array of states,
and the predict method returns a 2D array of action values
by the number of samples passed to the predict method.

Although we wrote this method to be able to predict multiple
output actions for multiple input states, we only ever really
need to pass one state. For that reason we limited the states
passed to one, and then assumming this to be the case we fixed
the parsing of the action out of the predicted Q values.