Changes to evaluation callback
- Add call on training start that evaluates performance of
  random policy and policy derived from initial model.
- Add calculation of metric measuring average number of
  timesteps per episode. Log that metric to Tensorboard.
- Move simulation logic into its own method so it can
  be shared by the on_epoch_end and on_train_start methods.

After doing this we found that the initial model (which is
initialized with random weights) performs worse than the
random policy, unless the initial weights are constant. This
suggests to us there is something wrong with the model since
the model shouldn't be biased towards a single at the beginning
if the initial weights are random. We are investigating this
further.