Create test runner
The test runner takes a model checkpoint, loads it into
the model, and then uses the agent methods to convert
the model into a policy. It also renders the simulation
of the policy being executed.

Using this script we found that the policy being learned
by our agent is biased towards the action 1. We are now
trying to determine why that is, whether its an issue with
the Keras model, with the EnvironmentSequence, or both.