Move model debugigng code to module, add weight/grad checks
We previously had a test at the end of the rundef file checking
whether the generator was working correctly. We have moved this
code into a function and put it into this module.

We also created a new function that takes in the model and the
generator and prints out the weights and gradients on different
layers of the model for that minibatch. We compare our outputs,
weights, and gradients calculated from the model to some calculated
by hand with TF.

Checking the gradients manually allowed us to solve a bug that
was causing issues with the gradient. Namely, the shape of the
output tensors both in the model and in the generator were incorrect.
This was causing the loss to be calculated incorrectly because
the incorrectly shaped tensors were being broadcast and combined
into the wrong shape. We've kept this method because of how
useful it is to spot check gradients.