Critical fix: Q_a output tensor shape fix (and other changes)
We were running into a major issue where the gradient was not
updating correctly. The value function would uniformly increase
towards 9.35, which is the value of the constant policy. It would
then stay there, and the losss would be constant and about 8. We
found that the Q function estimator was not fitting the "target"
value function at terminal states, and this was the source of the
error.

Because of this we suspected there was an error with the gradients.
We wrote a debugging function to print out the gradients and losses
for the network. We found that the loss had the wrong shape, and
the resulting loss and gradients were wrong. We found that the shape
problem was because the Q_a output (output 2 of the network) had the
wrong shape, so we fixed the shape of it in the model and also
in the generator. The changes to the genrator are in a previous
commit, and the changes to the model can be found under the
`get_Q_a` method in this file.

This solved the problem and our simple DQN for the cartpole problem
works now!!! Finally!!! :,)

Other changes we did along the way:

- Add aforementioned debugging functions to main method of this
  file.
- Initialize weights with random Gaussian distribution to avoid
  symmetry problems with weight values and gradient.
- Change model to use simple one that was found to work by
  someoene else (https://towardsdatascience.com/cartpole-introduction-to-reinforcement-learning-ed0eb5b58288#f94f)
  We wanted to make sure we had a simple but powerful enough
  model that we knew should work.
- Change optimizer to RMSProp, which is what was used in
  Atari paper. We thought this made sense since Adam is
  highly adaptive but adaptiveness and momentum may not
  make sense when the target model is being updated often.
- Make training policy more exploratory, increasing
  frequency of exploratory action from 0.05 to 0.1.
- Add `replay_buffer_min` parameter which controls how
  many iterations are used to prepopulate the replay
  memory with the random policy before value function
  learning starts.
- Other changes to hyperparameters to improve learning
  stability and speed, and model clarity:
    - Target frequency written in terms of minibatch
      updates to align more clearly with minibatches.
    - Eliminate action repeat, since episodes are short
      anyways so changing actions at each step is important.
    - Increase replay buffer size for stability, write in
      terms of minibatch size.
    - Increase number of episodes and initial states sampled
      to estimate model performance, to reduce variance of
      performance estimates.

- Add parameters that define which model checkpoint is used
  in `runners/test.py` for testing the model.

This commit marks the first, truly working version of this
repo!!! Yay!!!