Move methods specific to classic control to new sequence class
We're moving towards creating child classes implementing different
feature creation functions. In order to do this we extracted the
relevant functions from the original `EnvironmentSequence` class
(now named `SynchronousSequence` for generality) into a new class.
We named the new class `EnvironmentSequence` to keep our code com-
patible with the agents we've already implemented for classic control.
We may rename it later.

See the notes in the PR for the methods that were moved and why.