Allow modification of raw reward from environment
In the Atari papers the actual reward from the game is
modified to +1, -1, or 0 depending on whether its zero
or not and its sign. This is so that the same agent can
learn to play different games irrespective of the scale
of the rewards.

In order to allow this functionality in our framework
we have wrapped the reward buffer with a `get_reward_by_index`
method that allows us to do an intermediate transformation
on the rewards before using them to compute the minibatch
labels. This is similar to how we wrapper the observation
buffer with the `get_feature_by_index` method to allow
transformations of observations into features. It is the
next natural step towards abstracting environment outputs
so they can be preprocessed before passing them to the agent.

We modified the parent environment file to implement the
abstract method `get_reward_by_index` and then we implemented
corresponding methods in each of the two child classes. For
the `EnvironmentSequence` class used for classical discrete
control, there is no transform, we just feed the raw rewards
to the agent as we have been doing.