Add skip frames parameter to SynchronousSequence class
The `skip_frames` parameter determines whether frames
that the environment produces during skip actions are
saved in the buffer (and thus used for training) or not.
After reading the DQN and double DQN papers more closely
we realized that during action repeats frames seemed to
not be stored in the replay memory. The paper says the
DQN agents "do not see" these frames. Thus we added a
mechanism to skip these frames altogether and not add
them to the memory.

For backwards compatibility purposes this parameter
is defaulted to `False` in the class constructor.