Add `skip_frame` parameter to Atari Seq, return TOTAL rewards
Since we will be skipping frames altogether from the training
process we thought it would make sense to add all of the rewards
for the frames skipped and adding them together into a single reward.
That way if rewards are created by the environment in skipped
frames the agent still knows about them. We implemented this
by modifying the `get_rewards_by_index` method to look at
all previous rewards in the action repeat window.