Rename observation buffer, preprocess features before saving to replay mem
We hypothesized that in the Atari case if we moved the image
luminencense extraction and rescaling to before the image was
saved to the replay memory that we would see significant compute
and memory savings. This was indeed the case: we were able to
reduce the epoch time from 41 minutes to 15 minutes! We thus
created a new `preprocess_observation` abstract method in the
parent SynchronousSequence class and apply transformations to
the observations to create the base features (pre-stacking)
before saving them to the observation buffer. Since the observation
buffer now stores features rather than observations we renamed it
to `feature_buffer` in all SynchronousSequence sub-classes and
in the class itself.