Manually set ALE frameskip, reframe iteration values based on this
We discovered by looking at the OpenAI Gym documentation that the
environment class skips frames by default and reuses the same
action as was already passed to it!!! This explains why the running
time for our application was so ridiculously long, because we
were doing frameskipping twice, once manually in our implementation
and once implicitly through the simulator. We set the frameskipping
value for the simulator explicitly, got rid of our manual frame
skipping for this model, and changed our epoch, iteration, and
exploration schedule parameters to train the correct number of
frames based on this change. We updated comments throughout the
file to clarify.

We also added a parameter to the AtariSequence for taking pairwise
maxes explicitly. Pairwise maxes slow down performance noticeably
but are not needed for all Atari games.