Replace replay memory lists with dequeus
We suspect that the replay memory's management is significantly
slowing down the training when the replay memory is large. We took
a look at the replay memory class in keras-rl for inspiration. We
were happy to see a lot of his implementation coincides with ours,
without us referencing it, hurray! However, the author uses deques
instead of lists for the replay memory. This is likely because:

1) They automatically manage fixed length constraints
2) The memory management portion of them is O(1) instead of O(n)

We're hoping the latter point will help speed up the performance
of our buffer.