Add references to default graph to sequence classes and agent methods
We were running into a problem when running the model training and
the sequence in different threads (i.e `workers=1`). Basically the
copy of the model being trained and the copy being used as a target
would run in different threads. Since the model in the sequence thread
did not share a graph with the model in the model thread this raised an
error, summarized here:

https://github.com/keras-team/keras/issues/5511

We found that by setting the default graph when compiling the original
model, passing that graph to the generator, and setting that same graph
in the "context" (using a with statement) when copying new target models
or evaluating the target model solved this issue. We believe this forces
the target model to use the same graph as the model being trained. This
is a solution we found in the github thread linked, and had to extend
to our use case.

We also made a minor update to the assertion checking iterations and
batch indices agree, to make debugging easier. We have made changes
in the rundef file to match these but those are not in this commit.

Making this change has cut the running time of the model by more
than half, even with double DQN and frame pair max on!!! It runs
now in 7-8 minutes per epoch!!! :D I'm super happy about this fix.