Check frame stack does not intersect episode end in action selection
When selecting actions we pass a stack of frames (our feature) to the
model. In particular we pass the most recent stack of frames (the
rightmost end) in the replay memory. If the action is selected early
in a new episode some of the frames in the rencent frame stack will
include frames from the previous episode. Thus if we stacked these
frames we would be passing frames from two different episodes to
the model. This clearly wouldn't make sense as a feature.

We therefore made a change to action selection: When the most
recent frame stack crosses an episode boundary we select a
random action instead of passing the frame stack to the model.
This avoids this issue. Furthermore, since frame stacks are
relatively small (4-5 frames) and episode starts are not that
common this should not affect the performance of the agent
very much.