Remove shuffle of sequence, increase workers from 0 to 1
As mentioned in previous commits we solved the issue which
limited us from running two threads: one for creating minibatches
and another for backpropagating. The problem had to do with copies
of the same model running on different threads.

We also removed the shuffling on minibatch indices. Since our
agent is stateful our minibatch generation only makes sense in
order. Hence the change.