Add logic to create initial evaluation states and serialize at init
We found that sampling 10,000 initial states at the end of each epoch
would take 20 minutes each time. This is more than 2x the length of an
epoch itself!

Instead of doing this each epoch we made to changes:

- Move the sampling of initial states to the construction of the
  custom callback. This way we only pay the cost of sampling the
  states once per training rather than once per epoch.

- Serialize sampled initial states and save them into a compress
  .npz file. This way we only incur the sampling cost once **per
  model** (possibly per environment) and we can have a consistent
  baseline of states for comparing the performance of different
  versions of the model (or different, but similar models with
  the same environment).

Doing this makes evaluation extremely fast, at the price of the
initial run to generate the sampled initial states.