Move evaluation sequence constructor method to matching sequence subclass
We noticed that it would make more sense to have the constructor for
evaluation sequence subclass instances to be included as part of the
class itself. We added a method `create_validation_instance` to each
subclass of `SynchronousSequence`. This method returns an instance
of the subclass with parameters chosen to be used for the validation
callback. The method takes the evaluation exploration schedule as a
parameter. This is what is expected by the (updated -- see next commits)
evaluation callback.

We also created the constructor for the evaluation of agents using the
`EnvironmentSequence` class, which we previously had not implemented.
This allows us to use the evaluation callback in its current implementation
with our classical control agents. Thus this makes our changes to the callback
compatible backwards with our previous work.